# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Native Call Center API",
    "version": "1.0.0",
    "summary": "REST and SSE surface of the AI-native call center.",
...
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Same-origin API mount; the SPA is served by the same process."
    }
  ],
...
  "paths": {
...
    "/calls/{callId}/recordings": {
      "get": {
        "operationId": "listCallRecordings",
        "tags": [
          "recordings"
        ],
        "summary": "A call's audio artifacts",
        "description": "Requires a signed-in user: supervisors may list any call's recordings, an agent only those of a call they were on.",
        "parameters": [
          {
            "name": "callId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The call's recordings, possibly empty.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "cookieSession": [
              "history:read:own"
            ]
          },
          {
            "apiKeyBearer": [
              "history:read:own"
            ]
          }
        ]
      }
    },
...
    "/calls/{callId}/reviews": {
      "get": {
        "operationId": "listCallReviews",
        "tags": [
          "recordings"
        ],
        "summary": "A call's quality reviews",
        "parameters": [
          {
            "name": "callId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The call's reviews, possibly empty.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityReviewList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "cookieSession": [
              "quality:review"
            ]
          },
          {
            "apiKeyBearer": [
              "quality:review"
            ]
          }
        ]
      }
    },
...
    "/recordings/{recordingId}/audio": {
      "get": {
        "operationId": "getRecordingAudio",
        "tags": [
          "recordings"
        ],
        "summary": "Stream one recording",
        "description": "Serves the audio with HTTP range support, so players can seek. Conditional requests are honoured. Requires a signed-in user: supervisors may hear any recording, an agent only the calls they were on.",
        "parameters": [
          {
            "name": "recordingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The full audio.",
            "content": {
              "audio/wav": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "206": {
            "description": "The requested byte range of the audio.",
            "content": {
              "audio/wav": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "cookieSession": [
              "history:read:own"
            ]
          },
          {
            "apiKeyBearer": [
              "history:read:own"
            ]
          }
        ]
      }
    },
    "/recordings/{recordingId}/reviews": {
      "post": {
        "operationId": "createRecordingReview",
        "tags": [
          "recordings"
        ],
        "summary": "Score a recording",
        "description": "Records a quality review against the recording, attributed to the subject who wrote it.\n\n**The one operation with no apiKeyBearer branch.** A review is a person's judgement of another person's work, and reviewer_id has to name somebody who can be asked about it; a key acting for an agent would attribute the judgement to the agent being judged. This is not the web application being privileged — a supervisor's session token reaches it from curl exactly as it does from the browser. Reading reviews (GET /calls/{callId}/reviews) has a key branch like everything else.",
        "security": [
          {
            "cookieSession": [
              "quality:review"
            ],
            "csrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "recordingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The stored review.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityReview"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aicc_session",
        "description": "Session cookie issued by POST /auth/login, for a person signed in at a browser. The cookie name is configurable (AICC_SESSION_COOKIE); aicc_session is the default. HttpOnly, SameSite=Lax. The scopes it carries are derived from the account's role at sign-in; the role itself never appears on the wire."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AICC-Csrf",
        "description": "Anti-CSRF header, required on every mutating request that authenticates with the session cookie. Any non-empty value passes; requiring a custom header forces a CORS preflight on a cross-origin caller. Missing header: 403 FORBIDDEN. It never accompanies apiKeyBearer — this header protects a credential the browser sends automatically, and a key is sent deliberately."
      },
      "apiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key, presented as `Authorization: Bearer <key>`, for a system with no browser. The key is the subject; it is not a person and never becomes one. To act for an agent — a CRM dialling for somebody on the floor — send `X-AICC-Agent-ID: <agentId>` alongside it. Without that header the subject acts for nobody, and an agent-scoped operation answers AGENT_REQUIRED. The header is refused outright on a session request (AGENT_IMPERSONATION_NOT_ALLOWED), whatever the account's role. A key carries exactly the scopes it was issued with; missing one is INSUFFICIENT_SCOPE."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request is malformed: unparseable body or an invalid identifier. Codes VALIDATION_FAILED, USER_DATA_TOO_LARGE.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "No session or the session expired. Codes SESSION_EXPIRED, INVALID_CREDENTIALS.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The identity lacks the role, the CSRF header is missing, the account is suspended, or the account is not an agent. Codes FORBIDDEN, USER_SUSPENDED, NOT_CALL_PARTY.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "The target does not exist. Codes NOT_FOUND, CALL_NOT_FOUND.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "An unexpected failure. Code INTERNAL.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorCode": {
        "type": "string",
        "description": "Machine-readable, translatable failure identifier. The frontend renders errors.<CODE>; the backend never localizes.",
        "enum": [
          "INVALID_CREDENTIALS",
          "SESSION_EXPIRED",
          "FORBIDDEN",
          "AGENT_REQUIRED",
          "AGENT_IMPERSONATION_NOT_ALLOWED",
          "INSUFFICIENT_SCOPE",
          "VALIDATION_FAILED",
          "USER_DATA_TOO_LARGE",
          "NOT_FOUND",
          "METHOD_NOT_ALLOWED",
          "CONFLICT",
          "EXTENSION_IN_USE",
          "EXTENSION_ASSIGNED_TO_AGENT",
          "LAST_ADMIN",
          "EXTENSION_POOL_EXHAUSTED",
          "AGENT_ALREADY_LOGGED_IN",
          "AGENT_NOT_LOGGED_IN",
          "AGENT_NOT_IN_WRAP_UP",
          "CALL_NOT_FOUND",
          "NOT_CALL_PARTY",
          "OPERATION_NOT_ALLOWED_FOR_CALL_TYPE",
          "USER_SUSPENDED",
          "SWITCH_DOWN",
          "STORAGE_DOWN",
          "RATE_LIMITED",
          "INTERNAL"
        ]
      },
      "Error": {
        "type": "object",
        "description": "The single error envelope body: an error code plus interpolation params. Message is diagnostic English, never shown to end users.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "params": {
            "type": "object",
            "description": "Interpolation parameters for the localized message, e.g. {\"field\": \"extensionNumber\"}.",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Recording": {
        "type": "object",
        "description": "One stored audio artifact of a call.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "callId": {
            "type": "string",
            "format": "uuid"
          },
          "backend": {
            "type": "string",
            "description": "Storage backend, e.g. fs or s3."
          },
          "bucket": {
            "type": "string"
          },
          "objectKey": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "durationSec": {
            "type": "integer"
          },
          "format": {
            "type": "string",
            "description": "Audio container, e.g. wav."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "callId",
          "backend",
          "objectKey",
          "sizeBytes",
          "durationSec",
          "format",
          "createdAt"
        ]
      },
      "RecordingList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Recording"
            }
          }
        },
        "required": [
          "items"
        ]
      },
...
      "QualityReview": {
        "type": "object",
        "description": "A reviewer's scoring of one recording.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "recordingId": {
            "type": "string",
            "format": "uuid"
          },
          "callId": {
            "type": "string",
            "format": "uuid"
          },
          "reviewerId": {
            "type": "string",
            "format": "uuid"
          },
          "scores": {
            "type": [
              "object",
              "null"
            ],
            "description": "Per-criterion scores; null when the reviewer submitted none.",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "totalScore": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "notes": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "recordingId",
          "callId",
          "reviewerId",
          "scores",
          "totalScore",
          "notes",
          "createdAt"
        ]
      },
      "QualityReviewList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QualityReview"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "CreateReviewRequest": {
        "type": "object",
        "properties": {
          "scores": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "totalScore": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
    }
  }
}